Refactor region and base name utilities into shared module
